+++ pkg/executor/nfs/nfsrestoreresources.go
@@ -4,19 +4,17 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"path/filepath"
 
 	"github.com/libopenstorage/stork/drivers/volume"
 	"github.com/libopenstorage/stork/drivers/volume/kdmp"
 	storkapi "github.com/libopenstorage/stork/pkg/apis/stork/v1alpha1"
-	"github.com/libopenstorage/stork/pkg/crypto"
 	"github.com/libopenstorage/stork/pkg/k8sutils"
 	"github.com/libopenstorage/stork/pkg/log"
 	kdmpapi "github.com/portworx/kdmp/pkg/apis/kdmp/v1alpha1"
 	"github.com/portworx/kdmp/pkg/drivers/utils"
 	"github.com/portworx/kdmp/pkg/executor"
-	"github.com/portworx/sched-ops/k8s/core"
+	//"github.com/portworx/sched-ops/k8s/core"
 	kdmpschedops "github.com/portworx/sched-ops/k8s/kdmp"
 	storkops "github.com/portworx/sched-ops/k8s/stork"
 	"github.com/sirupsen/logrus"
@@ -57,7 +55,7 @@ func restoreAndApplyResources(
 	rbCrNamespace string,
 ) error {
 
-	err := createNamespacesFromMapping(applicationCRName, restoreNamespace)
+	err := executor.CreateNamespacesFromMapping(applicationCRName, restoreNamespace)
 	if err != nil {
 		//update resourcebackup CR with status and reason
 		logrus.Errorf("restore resources for [%v/%v] failed with error: %v", rbCrNamespace, rbCrName, err.Error())
@@ -128,7 +126,6 @@ func restoreResources(
 		return err
 	}
 	objects, err := downloadResources(backup, restore.Spec.BackupLocation, restore.Namespace)
-	//objects, err := downloadResources(backup, restore)
 	if err != nil {
 		log.ApplicationRestoreLog(restore).Errorf("Error downloading resources: %v", err)
 		return err
@@ -147,7 +144,7 @@ func downloadCRD(
 ) error {
 	var crds []*apiextensionsv1beta1.CustomResourceDefinition
 	var crdsV1 []*apiextensionsv1.CustomResourceDefinition
-	crdData, err := downloadObject(resourcePath, resourceFileName, encryptionKey)
+	crdData, err := executor.DownloadObject(resourcePath, resourceFileName, encryptionKey)
 	if err != nil {
 		return err
 	}
@@ -227,29 +224,6 @@ func downloadCRD(
 	return nil
 }
 
-func downloadObject(
-	resourcePath string,
-	resourceFileName string,
-	encryptionKey string,
-) ([]byte, error) {
-	logrus.Debugf("downloadObject resourcePath: %s, resourceFileName: %s", resourcePath, resourceFileName)
-	filePath := filepath.Join(resourcePath, resourceFileName)
-	data, err := ioutil.ReadFile(filePath)
-	if err != nil {
-		return nil, fmt.Errorf("getting file content of %s failed: %v", filePath, err)
-	}
-
-	if len(encryptionKey) > 0 {
-		var decryptData []byte
-		if decryptData, err = crypto.Decrypt(data, encryptionKey); err != nil {
-			logrus.Errorf("nfs downloadObject: decrypt failed :%v, returning data direclty", err)
-			return data, nil
-		}
-		return decryptData, nil
-	}
-	return data, nil
-}
-
 func downloadResources(
 	backup *storkapi.ApplicationBackup,
 	backupLocation string,
@@ -271,7 +245,7 @@ func downloadResources(
 	if err := downloadCRD(bkpDir, crdFile, restoreLocation.Location.EncryptionV2Key); err != nil {
 		return nil, fmt.Errorf("error downloading CRDs: %v", err)
 	}
-	data, err := downloadObject(bkpDir, resourcesFile, restoreLocation.Location.EncryptionV2Key)
+	data, err := executor.DownloadObject(bkpDir, resourcesFile, restoreLocation.Location.EncryptionV2Key)
 	if err != nil {
 		return nil, fmt.Errorf("error downloading resources: %v", err)
 	}
@@ -630,134 +604,4 @@ func getDynamicInterface() (dynamic.Interface, error) {
 	}
 
 	return dynamic.NewForConfig(config)
-}
-
-func createNamespacesFromMapping(
-	applicationCRName string,
-	restoreNamespace string) error {
-	funct := "createNamespacesFromMapping"
-	restore, err := storkops.Instance().GetApplicationRestore(applicationCRName, restoreNamespace)
-	if err != nil {
-		logrus.Errorf("%s: Error getting restore cr: %v", funct, err)
-		return err
-	}
-	backup, err := storkops.Instance().GetApplicationBackup(restore.Spec.BackupName, restore.Namespace)
-	if err != nil {
-		log.ApplicationRestoreLog(restore).Errorf("Error getting backup: %v", err)
-		return err
-	}
-	return createNamespaces(backup, restore.Spec.BackupLocation, restore.Namespace, restore)
-}
-
-func createNamespaces(backup *storkapi.ApplicationBackup,
-	backupLocation string,
-	backupLocationNamespace string,
-	restore *storkapi.ApplicationRestore) error {
-	var namespaces []*v1.Namespace
-	funct := "create namespaces"
-
-	repo, err := executor.ParseCloudCred()
-	if err != nil {
-		logrus.Errorf("%s: error parsing cloud cred: %v", funct, err)
-		return err
-	}
-	bkpDir := filepath.Join(repo.Path, backup.Status.BackupPath)
-	restoreLocation, err := storkops.Instance().GetBackupLocation(backup.Spec.BackupLocation, backupLocationNamespace)
-	if err != nil {
-		return err
-	}
-
-	nsData, err := downloadObject(bkpDir, namespacesFile, restoreLocation.Location.EncryptionKey)
-	if err != nil {
-		return err
-	}
-	if nsData != nil {
-		if err = json.Unmarshal(nsData, &namespaces); err != nil {
-			return err
-		}
-		for _, ns := range namespaces {
-			if restoreNS, ok := restore.Spec.NamespaceMapping[ns.Name]; ok {
-				ns.Name = restoreNS
-			} else {
-				// Skip namespaces we aren't restoring
-				continue
-			}
-			// create mapped restore namespace with metadata of backed up
-			// namespace
-			_, err := core.Instance().CreateNamespace(&v1.Namespace{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:        ns.Name,
-					Labels:      ns.Labels,
-					Annotations: ns.GetAnnotations(),
-				},
-			})
-			log.ApplicationRestoreLog(restore).Tracef("Creating dest namespace %v", ns.Name)
-			if err != nil {
-				if errors.IsAlreadyExists(err) {
-					oldNS, err := core.Instance().GetNamespace(ns.GetName())
-					if err != nil {
-						return err
-					}
-					annotations := make(map[string]string)
-					labels := make(map[string]string)
-					if restore.Spec.ReplacePolicy == storkapi.ApplicationRestoreReplacePolicyDelete {
-						// overwrite all annotation in case of replace policy set to delete
-						annotations = ns.GetAnnotations()
-						labels = ns.GetLabels()
-					} else if restore.Spec.ReplacePolicy == storkapi.ApplicationRestoreReplacePolicyRetain {
-						// only add new annotation,labels in case of replace policy is set to retain
-						annotations = oldNS.GetAnnotations()
-						if annotations == nil {
-							annotations = make(map[string]string)
-						}
-						for k, v := range ns.GetAnnotations() {
-							if _, ok := annotations[k]; !ok {
-								annotations[k] = v
-							}
-						}
-						labels = oldNS.GetLabels()
-						if labels == nil {
-							labels = make(map[string]string)
-						}
-						for k, v := range ns.GetLabels() {
-							if _, ok := labels[k]; !ok {
-								labels[k] = v
-							}
-						}
-					}
-					log.ApplicationRestoreLog(restore).Tracef("Namespace already exists, updating dest namespace %v", ns.Name)
-					_, err = core.Instance().UpdateNamespace(&v1.Namespace{
-						ObjectMeta: metav1.ObjectMeta{
-							Name:        ns.Name,
-							Labels:      labels,
-							Annotations: annotations,
-						},
-					})
-					if err != nil {
-						return err
-					}
-					continue
-				}
-				return err
-			}
-		}
-		return nil
-	}
-	for _, namespace := range restore.Spec.NamespaceMapping {
-		if ns, err := core.Instance().GetNamespace(namespace); err != nil {
-			if errors.IsNotFound(err) {
-				if _, err := core.Instance().CreateNamespace(&v1.Namespace{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:        ns.Name,
-						Labels:      ns.Labels,
-						Annotations: ns.GetAnnotations(),
-					},
-				}); err != nil {
-					return err
-				}
-			}
-			return err
-		}
-	}
-	return nil
 }
